Ready to merge. added deleteRating functionality to rating controller/service/repisotory

See merge request cohorts/usaa-efit-vetfit/usaa-efit-jan-2023/student-work/capstone-project/blues-twos-movie-reviews!12

+++ src/main/java/com/galvanize/bluestwosmoviereviews/services/RatingService.java
@@ -39,8 +39,14 @@ public class RatingService {
         return ratingRepository.findByUserId(userId);
     }
 
-    public void deleteById(Integer ratingId) {
-        ratingRepository.deleteById(ratingId);
+    public RatingModel deleteById(Integer ratingId) {
+        RatingModel ratingToDelete = ratingRepository.findByRatingId(ratingId);
+
+        if (ratingId != null) {
+            ratingRepository.deleteById(ratingId);
+        }
+
+        return ratingToDelete;
     }
 
     public RatingModel addNewRating(RatingModel ratingModel) {
